Hi Baris,
I set up a basic structure for the rendering and implemented some graphical objects.
At the moment we have a board and the pegs. We can render them and move them on the board.
It is just a basic board. Until now, we have no textures. The pegs are cuboids (like blocks).
I copied some pieces from the project I gave to you but I changed all the structure to a better one, where we can easyly add more stuff.
I added some team identity to the player and the pegs in order to manage them.
I created most of the structure in two steps, meaning that we have a basic class describing a peg or a board and then we extend those
for implementing a specific version, like a classic board or a simple peg.
Lots of things are static now, because it's faster and easier for a mobile device and for the total project as well.

Some of the next steps might be:
- adding a bluetooth connection
- making the pegs moving slowly and not just jumping to the next field
- completing the methods of the board
- implementing the player and the dice
- setting up a basic GUI for controlling the game including some menus for settings and the possibility to enable and disable certain features and animations
- changing the camera settings, so that the player can control the view and rotate the board (maybe with phone acceleration)
- adding textures for a better board and creating a more realistic peg
- adding lots of features like 'shaking for throwing the dice', ...

Important:
I tested the game on the emulator and it runs, but the graphics sometimes lags and you can see an incomplete rendering
and some triangles are rendered only partly.
But I tested it also on a real device and there is no single mistake :)

So don't be surprised, if the emulator doesn't do it perfectly ..

However,
feel free to choose any subject - just tell me when you start with something, so that we don't work on the same thing ;)

So long,
Stefan