Hi Baris,
I improved the board and the peg's behavior as well as the players'.
Added a dice for rendering. Now dice calculation is working without interrupting the game cycle :)
Also dice is rotating and moving with OpenGL. I will add a better moves for the dice, too.

Added an option for highlighting a peg (maybe not the best one ;)
What do you think about the idea to show a highlighted (selected) peg still with its color, but on some bright, blinking animated square ?
(Square would be between the peg's bottom and the board all the time and only visible and painted when the peg is selected. I think that might look better :)

I'm going to work on the user interface now,
so that the player is asked to throw the dice and to confirm the current selection ..

How is it going with the bluetooth connection ?

So long, Stefan