Okay :)
I changed the game cycle and and the way to keep it alive.
Now the actions are called by one of the notifications.
So basically, we don't do anything about the multiplayer-mode and the objects don't know whether it is enabled or not.
The only use of the NetworkPlayer is to stop the game cycle, but to make this possible, the cycle is running on all devices.
That means, that after every action still a notification is sent. But now it's subject to this notification to make the game keep running.
I just tested it with three devices. The dice and the peg movement is transferred to both devices successful.
Anyway, we don't have a queue for the events, yet
and we can't sent notifications back to the server. So this might be the next thing to do.
Also the game is started right after we press the multiplayer-button - actually the clients should ask the server about the game.
Sometimes I get a socket closed error, but the game keeps running.
I'm working on it. The next thing I will do is implementing a queue for the notifications.

So long, Stefan